+++ src/migrations/Install.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * Template Select plugin for Craft CMS 5.x
+ *
+ * A fieldtype that allows you to select a template from a dropdown.
+ *
+ * @link      https://superbig.co
+ * @copyright Copyright (c) 2017 Superbig
+ */
+
+namespace superbig\templateselect\migrations;
+
+use craft\db\Migration;
+
+/**
+ * Install migration.
+ */
+class Install extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function safeUp(): bool
+    {
+        // Nothing to install - field type is registered via event listener
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown(): bool
+    {
+        return true;
+    }
+}